chore: adapt code to upgraded scoringrules, ruff and pandas-stubs
scoringrules 0.10:
- crps test: drop quantile levels 0.0/1.0 (0.10 rejects them; they
  contribute zero pinball loss, so the value becomes 8/9 not 8/11).
- rcrps tests: give one forecast row per observation (0.10 no longer
  broadcasts a single row); expected values unchanged.

ruff 0.15 (new default rules):
- RUF069: constant_component_splitter now uses math.isclose for the
  ratio-sum check (genuine float-fragility fix); the plotter compares
  integer percentiles; data_split keeps an exact-sentinel noqa.
- D420: reorder Raises after Args/Returns in testing.py.
- PLW0717: justified noqa on two intentional broad pipeline try-blocks.
- Formatter reformats a few method chains / conditional expressions.

pandas-stubs 2.3.3 (pyright):
- replace functools.reduce(iadd) with a comprehension, drop a now-
  unnecessary cast, add casts/ignores for stub-strictness changes.

+++ packages/openstef-beam/src/openstef_beam/analysis/plots/forecast_time_series_plotter.py
@@ -396,7 +396,7 @@ class ForecastTimeSeriesPlotter(BaseConfig):
             # Create band data from widest to narrowest
             for i in range(len(percentiles) // 2):
                 lower_quantile, upper_quantile = percentiles[i], percentiles[-(i + 1)]
-                if float(lower_quantile) == self.MEDIAN_QUANTILE:
+                if lower_quantile == int(self.MEDIAN_QUANTILE):
                     continue
 
                 # type: ignore construct a BandData-compatible dict
